stream: factor out Flow methods to traits
This will allow other libraries to more easily incorporate the Jelly Pekko Streams APIs into their design.

The integration tests seem to be broken, but I'm not sure if the changes here are the reason.
